feat: Usable Bar chart
usable bar chart and some fix in tab_overview
